JPA CRUD demo - delete tất cả record bằng method createQuery của Entity Manager

+++ src/main/java/com/luv2code/cruddemo/dao/StudentDAOImpl.java
@@ -54,6 +54,14 @@ public class StudentDAOImpl implements StudentDAO {
     @Transactional
     public void delete(Integer id) {
         Student theStudent = entityManager.find(Student.class, id);
-        entityManager.remove(theStudent);
+        if(theStudent!=null)
+            entityManager.remove(theStudent);
+    }
+
+    @Override
+    @Transactional
+    public int deleteAll() {
+        int numRowsDeleted = entityManager.createQuery("delete from Student ").executeUpdate();
+        return numRowsDeleted;
     }
 }
